Fixed typo 'insurence' to 'insurance'
