+++ template/src/main/java/pot/insurance/manager/repository/CompanyRepository.java
@@ -1,0 +1,10 @@
+package pot.insurance.manager.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import pot.insurance.manager.domain.Company;
+
+import java.util.UUID;
+
+public interface CompanyRepository extends CrudRepository<Company, UUID> {
+
+}
