+++ template/src/main/java/pot/insurance/manager/service/CompanyService.java
@@ -1,10 +1,10 @@
-package pot.insurence.manager.service;
+package pot.insurance.manager.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import pot.insurence.manager.domain.Company;
-import pot.insurence.manager.repository.CompanyRepository;
-import pot.insurence.manager.dto.CompanyDTO;
+import pot.insurance.manager.domain.Company;
+import pot.insurance.manager.dto.CompanyDTO;
+import pot.insurance.manager.repository.CompanyRepository;
 
 import java.util.UUID;
